Refactor navigation tests; move navigation tests to a new file for improved organization and maintainability
