+++ cypress/support/navigation/navigation-to-user-dashboard.ts
